+++ src/Models/Taggable.php
@@ -11,39 +11,34 @@ class Taggable extends Model
 {
     use SoftDeletes;
 
-    private $uuids = false;
+    private bool $uuids, $taggableUuids = false;
+    private string $fillablesTagId = 'tag_id';
+    private string $fillablesTaggableId = 'taggable_id';
+    public array $fillables = ['taggable_type'];
 
     public function __construct(array $attributes = [])
     {
         parent::__construct($attributes);
         $this->setConnection(config('briofy-tag.database.connection'));
-        $this->uuids = config('briofy-tag.database.taggable_uuid', false);
-        if($this->uuids){
+        $this->uuids = config('briofy-tag.database.uuid', false);
+        $this->taggableUuids = config('briofy-tag.database.taggable_uuid', false);
+        if($this->taggableUuids){
             $this->primaryKey = 'uuid';
             $this->keyType = 'string';
             $this->incrementing = false;
+            $this->fillablesTaggableId = 'taggable_uuid';
         }
+        if($this->uuids) $this->fillablesTagId = 'tag_uuid';
+        $this->fillables = array_merge($this->fillables, [$this->fillablesTagId, $this->fillablesTaggableId]);
     }
 
-    protected $fillable = [ 'tag_id', 'taggable_id', 'taggable_type'];
-
-    protected function initializeTraits()
+    /**
+     * @param string[] $fillable
+     */
+    public function setFillable(array $fillable): void
     {
-        parent::initializeTraits();
-        if($this->uuids){
-            $this->bootHasUuids();
-        }
+        $this->fillable = $this->fillables;
     }
-
-    protected static function bootHasUuids()
-    {
-        static::creating(function ($model) {
-            if (!$model->getKey()) {
-                $model->{$model->getKeyName()} = (string) Str::orderedUuid();
-            }
-        });
-    }
-
     public function tag()
     {
         return $this->belongsTo(Tag::class);